chore(node): declutter global definitions
