+++ types/depd/index.d.ts
@@ -26,16 +26,14 @@ declare namespace depd {
     }
 }
 
-declare global {
-    namespace NodeJS {
-        interface Process {
-            addListener(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
-            emit(event: 'deprecation', code: depd.DeprecationError): boolean;
-            on(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
-            once(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
-            prependListener(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
-            prependOnceListener(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
-            listeners(event: 'deprecation'): depd.DeprecationError[];
-        }
+declare module 'process' {
+    interface Process {
+        addListener(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
+        emit(event: 'deprecation', code: depd.DeprecationError): boolean;
+        on(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
+        once(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
+        prependListener(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
+        prependOnceListener(event: 'deprecation', listener: (deprecationError: depd.DeprecationError) => void): this;
+        listeners(event: 'deprecation'): depd.DeprecationError[];
     }
 }
