+++ types/gulp-zip/index.d.ts
@@ -27,6 +27,6 @@ declare namespace GulpZip {
     }
 }
 
-declare function GulpZip(filename: string, options?: GulpZip.GulpZipOptions): NodeJS.ReadStream;
+declare function GulpZip(filename: string, options?: GulpZip.GulpZipOptions): NodeJS.ReadableStream;
 
 export = GulpZip;
