--- types/net-keepalive/tslint.json
@@ -1,12 +1,0 @@
-{
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "eofline": false,
-        "indent": [
-            true,
-            "spaces",
-            4
-        ],
-        "semicolon": false
-    }
-}
