+++ types/node/events.d.ts
@@ -35,9 +35,6 @@ declare module "events" {
          */
         let captureRejections: boolean;
 
-        interface EventEmitter extends NodeJS.EventEmitter {
-        }
-
         class EventEmitter {
             constructor(options?: EventEmitterOptions);
             /** @deprecated since v4.0.0 */
@@ -53,6 +50,30 @@ declare module "events" {
              * regular `'error'` listener is installed.
              */
             static readonly errorMonitor: unique symbol;
+
+            addListener(event: string | symbol, listener: (...args: any[]) => void): this;
+            on(event: string | symbol, listener: (...args: any[]) => void): this;
+            once(event: string | symbol, listener: (...args: any[]) => void): this;
+            removeListener(event: string | symbol, listener: (...args: any[]) => void): this;
+            off(event: string | symbol, listener: (...args: any[]) => void): this;
+            removeAllListeners(event?: string | symbol): this;
+            setMaxListeners(n: number): this;
+            getMaxListeners(): number;
+            listeners(event: string | symbol): Function[];
+            rawListeners(event: string | symbol): Function[];
+            emit(event: string | symbol, ...args: any[]): boolean;
+            listenerCount(type: string | symbol): number;
+            // Added in Node 6...
+            prependListener(event: string | symbol, listener: (...args: any[]) => void): this;
+            prependOnceListener(event: string | symbol, listener: (...args: any[]) => void): this;
+            eventNames(): Array<string | symbol>;
+        }
+    }
+
+    global {
+        namespace NodeJS {
+            interface EventEmitter extends EventEmitter.EventEmitter {
+            }
         }
     }
