+++ types/node/stream.d.ts
@@ -344,5 +344,32 @@ declare module "stream" {
         }
     }
 
+    global {
+        namespace NodeJS {
+            interface ReadableStream extends events.EventEmitter {
+                readable: boolean;
+                read(size?: number): string | Buffer;
+                setEncoding(encoding: BufferEncoding): this;
+                pause(): this;
+                resume(): this;
+                isPaused(): boolean;
+                pipe<T extends WritableStream>(destination: T, options?: { end?: boolean; }): T;
+                unpipe(destination?: WritableStream): this;
+                unshift(chunk: string | Uint8Array, encoding?: BufferEncoding): void;
+                wrap(oldStream: ReadableStream): this;
+                [Symbol.asyncIterator](): AsyncIterableIterator<string | Buffer>;
+            }
+
+            interface WritableStream extends events.EventEmitter {
+                writable: boolean;
+                write(buffer: Uint8Array | string, cb?: (err?: Error | null) => void): boolean;
+                write(str: string, encoding?: BufferEncoding, cb?: (err?: Error | null) => void): boolean;
+                end(cb?: () => void): void;
+                end(data: string | Uint8Array, cb?: () => void): void;
+                end(str: string, encoding?: BufferEncoding, cb?: () => void): void;
+            }
+        }
+    }
+
     export = internal;
 }
