+++ types/node/ts3.2/base.d.ts
@@ -20,3 +20,4 @@
 /// <reference path="fs.d.ts" />
 /// <reference path="util.d.ts" />
 /// <reference path="globals.d.ts" />
+/// <reference path="process.d.ts" />
