+++ types/node/ts3.2/process.d.ts
@@ -1,0 +1,6 @@
+import 'process';
+declare module 'process' {
+    interface HRTime {
+        bigint(): bigint;
+    }
+}
