+++ types/undertaker/index.d.ts
@@ -9,6 +9,7 @@
 import * as Registry from "undertaker-registry";
 import { Duplex } from "stream";
 import { EventEmitter } from "events";
+import { Process } from 'process';
 
 declare namespace Undertaker {
     interface TaskFunctionParams {
@@ -23,7 +24,7 @@ declare namespace Undertaker {
     }
 
     interface TaskFunctionBase {
-        (done: (error?: any) => void): void | Duplex | NodeJS.Process | Promise<never> | any;
+        (done: (error?: any) => void): void | Duplex | Process | Promise<never> | any;
     }
 
     interface TaskFunction extends TaskFunctionBase, TaskFunctionParams {}
